% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjdproduction-package.R
\docType{package}
\name{rjdproduction-package}
\alias{rjdproduction}
\alias{rjdproduction-package}
\title{rjdproduction: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Tanguy Barthelemy \email{tanguy.barthelemy@insee.fr}

}
\keyword{internal}
